Move `TokenStringOidcClient` to `sigstore-cli` package
