+++ sigstore-cli/build.gradle.kts
@@ -12,6 +12,9 @@ dependencies {
     implementation("info.picocli:picocli:4.7.3")
     implementation("com.google.guava:guava:31.1-jre")
 
+    implementation(platform("com.google.oauth-client:google-oauth-client-bom:1.34.1"))
+    implementation("com.google.oauth-client:google-oauth-client")
+
     annotationProcessor("info.picocli:picocli-codegen:4.7.3")
 }
