+++ sigstore-cli/src/main/java/dev/sigstore/cli/TokenStringOidcClient.java
@@ -13,10 +13,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package dev.sigstore.oidc.client;
+package dev.sigstore.cli;
 
 import com.google.api.client.json.gson.GsonFactory;
 import com.google.api.client.json.webtoken.JsonWebSignature;
+import dev.sigstore.oidc.client.ImmutableOidcToken;
+import dev.sigstore.oidc.client.OidcClient;
+import dev.sigstore.oidc.client.OidcException;
+import dev.sigstore.oidc.client.OidcToken;
 import java.io.IOException;
 
 public class TokenStringOidcClient implements OidcClient {
